The restructured code
